standalone_modules/mockplacement/requirements.txt
fastapi
uvicorn
pymongo
email-validator
bcrypt
python-jose[cryptography]
python-multipart
python-dotenv
pydantic-settings
dnspython
